misc updates

+++ startup.sh
@@ -16,17 +16,19 @@ tail -f $FIFO | wob -a bottom \
 
 
 # Things after this are only run on a TTY
-[[ -n "$QTILE_XEPHYR" ]] && exit  
+[[ -n "$QTILE_XEPHYR" ]] || exit  
 
 
-run_if_new() { pgrep $1 &> /dev/null || $@ &; }
+run_if_new() { pgrep $1 || $@; }
 
-run_if_new mako
-run_if_new keepassxc
+{
+    run_if_new mako &
+    run_if_new keepassxc &
 
-pgrep wlsunset && pkill -SIGKILL wlsunset
-wlsunset -t 3100 -T 5700 -l 55.7 -L -3.1 &> /dev/null
+    pgrep wlsunset && pkill -SIGKILL wlsunset
+    wlsunset -t 3100 -T 5700 -l 55.7 -L -3.1 &
+    kanshi &
+} &> /dev/null
 
-sleep 5
+sleep 3
 check_systemd
-kanshi &> /dev/null &
